feat: cria endpoint para favoritar e desfavoritar uma receita

+++ app/Controllers/Http/FavoritoController.js
@@ -1,42 +1,45 @@
 const Favorito = use('App/Models/Favorito');
 
 class FavoritoController {
-  async index({ request }) {
-    const { page, perPage } = request.all();
+  async index({ auth, request }) {
+    const usuario_id = auth.user.id;
+    const { page, perPage = 10 } = request.all();
 
-    return Favorito.query().paginate(page, perPage);
+    return Favorito.query()
+      .select('id', 'usuario_id', 'receita_id')
+      .where({ usuario_id })
+      .with('receita')
+      .paginate(page, perPage);
   }
 
-  async store({ request }) {
-    const campos = Favorito.getCamposCadastro();
-    const dados = request.only(campos);
+  async favoritar({ auth, request }) {
+    const usuario_id = auth.user.id;
+    const { is_favorito, receita_id } = request.only([
+      'is_favorito',
+      'receita_id',
+    ]);
 
-    return await Favorito.create(dados);
-  }
+    const favorito = await Favorito.query()
+      .where({ usuario_id, receita_id })
+      .first();
 
-  async show({ params }) {
-    return await Favorito.query().where('id', params.id).first();
-  }
+    if (favorito) {
+      if (is_favorito === false) {
+        await favorito.delete();
 
-  async update({ params, request, response }) {
-    const favorito = await Favorito.findOrFail(params.id);
+        return {
+          mensagem: 'Receita desfavoritada com sucesso.',
+        };
+      }
 
-    const campos = Favorito.getCamposCadastro();
-    const dados = request.only(campos);
+      return null;
+    } else {
+      if (is_favorito) {
+        return await Favorito.create({ receita_id, usuario_id });
+      }
 
-    favorito.merge(dados);
-    await favorito.save();
-    return favorito;
-  }
-
-  async destroy({ params }) {
-    const favorito = await Favorito.findOrFail(params.id);
-
-    await favorito.delete();
-
-    return {
-      mensagem: 'Receita desfavoritada com sucesso.',
-    };
+      return null;
+    }
   }
 }
